Clarify data types of mountain_peak

+++ layers/mountain_peak/layer.sql
@@ -1,11 +1,11 @@
 CREATE OR REPLACE FUNCTION layer_mountain_peak(bbox geometry, zoom_level integer, pixel_width numeric)
 RETURNS TABLE(osm_id bigint, geometry geometry, name text, name_en text, ele int, ele_ft int, "rank" int) AS $$
    -- etldoc: osm_housenumber_point -> layer_housenumber:z14_
-   SELECT osm_id, geometry, name, name_en, ele, ele_ft, rank
+   SELECT osm_id, geometry, name, name_en, ele::int, ele_ft::int, rank::int
    FROM (
      SELECT osm_id, geometry, name, name_en,
      substring(ele from E'^(-?\\d+)(\\D|$)')::int AS ele,
-     round(substring(ele from E'^(-?\\d+)(\\D|$)')::int*3.2808399) AS ele_ft,
+     round(substring(ele from E'^(-?\\d+)(\\D|$)')::int*3.2808399)::int AS ele_ft,
        row_number() OVER (
            PARTITION BY LabelGrid(geometry, 100 * pixel_width)
            ORDER BY (
